@@ -8,7 +8,7 @@
         "start": "ng serve",
         "dev-aot": "ng serve --aot",
         "build": "ng build",
-        "build:prod": "ng build --target=production --environment=prod",
+        "build:prod": "ng build --prod",
         "test": "ng test",
         "lint": "ng lint",
         "e2e": "ng e2e",
@@ -19,51 +19,44 @@
     "author": "",
     "license": "EPL-2.0 OR Apache-2.0",
     "dependencies": {
-        "@angular/animations": "4.4.7",
-        "@angular/common": "4.4.7",
-        "@angular/compiler": "4.4.7",
-        "@angular/core": "4.4.7",
-        "@angular/forms": "4.4.7",
-        "@angular/http": "4.4.7",
-        "@angular/platform-browser": "4.4.7",
-        "@angular/platform-browser-dynamic": "4.4.7",
-        "@angular/router": "4.4.7",
+        "@angular/animations": "6.0.6",
+        "@angular/common": "6.0.6",
+        "@angular/compiler": "6.0.6",
+        "@angular/core": "6.0.6",
+        "@angular/forms": "6.0.6",
+        "@angular/http": "6.0.6",
+        "@angular/platform-browser": "6.0.6",
+        "@angular/platform-browser-dynamic": "6.0.6",
+        "@angular/router": "6.0.6",
         "angular2-websocket": "0.9.6",
-        "core-js": "2.5.4",
+        "core-js": "2.5.7",
         "font-awesome": "4.7.0",
-        "ngx-chips": "1.5.11",
         "ng2-file-upload": "1.3.0",
         "ng2-select": "2.0.0",
         "ng2-table": "1.3.2",
-        "ng2-toastr": "4.1.2",
+        "ngx-toastr": "8.8.0",
         "ngx-bootstrap": "2.0.3",
-        "ngx-pagination": "3.1.0",
+        "ngx-chips": "1.9.2",
+        "ngx-pagination": "3.1.1",
         "reflect-metadata": "0.1.12",
-        "rxjs": "5.5.8",
+        "rxjs": "6.2.0",
+        "rxjs-compat": "6.2.0",
         "spinkit": "1.2.5",
-        "zone.js": "0.8.25"
+        "zone.js": "0.8.26"
     },
     "devDependencies": {
-        "@angular/cli": "1.6.6",
-        "@angular/compiler-cli": "4.4.7",
-        "@angular/language-service": "4.4.7",
-        "@types/jasmine": "2.8.6",
+        "@angular/cli": "6.0.8",
+        "@angular/compiler-cli": "6.0.6",
+        "@angular-devkit/build-angular": "0.6.8",
+        "@angular/language-service": "6.0.6",
+        "@types/jasmine": "2.8.8",
         "@types/jasminewd2": "2.0.3",
-        "@types/node": "9.6.2",
-        "codelyzer": "3.1.2",
-        "file-loader": "^0.11.2",
-        "jasmine-core": "2.7.0",
-        "jasmine-spec-reporter": "4.2.1",
-        "karma": "1.7.0",
-        "karma-chrome-launcher": "2.2.0",
-        "karma-cli": "1.0.1",
-        "karma-coverage-istanbul-reporter": "1.3.0",
-        "karma-jasmine": "1.1.0",
-        "karma-jasmine-html-reporter": "0.2.2",
-        "protractor": "5.1.2",
-        "ts-node": "3.3.0",
-        "tslint": "5.6.0",
+        "@types/node": "10.3.2",
+        "codelyzer": "4.3.0",
+        "protractor": "5.3.2",
+        "ts-node": "5.0.1",
+        "tslint": "5.10.0",
         "typedoc": "0.11.1",
-        "typescript": "2.4.2"
+        "typescript": "2.7.2"
     }
 }
