@@ -1,69 +1,53 @@
 {
     "name": "workflow-modeler",
-    "version": "2.0.0",
+    "version": "2.0.0-SNAPSHOT",
     "description": "A designer for BPMN4TOSCA-based workflows",
     "scripts": {
         "start": "ng serve --port 9527",
         "dev-aot": "ng serve --aot --port 9527",
         "test": "ng test",
         "build": "ng build",
-        "build:prod": "ng build --target=production --environment=prod",
+        "build:prod": "ng build --prod",
         "lint": "ng lint"
     },
-    "licenses": [
-        {
-            "type": "Apache-2.0",
-            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
-        },
-        {
-            "type": "Eclipse Public License-2.0",
-            "url": "http://www.eclipse.org/legal/epl-v20.html"
-        }
-    ],
+    "license": "EPL-2.0 OR Apache-2.0",
     "dependencies": {
-        "@angular/animations": "4.3.5",
-        "@angular/common": "4.3.5",
-        "@angular/compiler": "4.3.5",
-        "@angular/core": "4.3.5",
-        "@angular/forms": "4.3.5",
-        "@angular/http": "4.3.5",
-        "@angular/platform-browser": "4.3.5",
-        "@angular/platform-browser-dynamic": "4.3.5",
-        "@angular/router": "4.3.5",
-        "angular-in-memory-web-api": "0.3.2",
+        "@angular/animations": "6.0.6",
+        "@angular/common": "6.0.6",
+        "@angular/compiler": "6.0.6",
+        "@angular/core": "6.0.6",
+        "@angular/forms": "6.0.6",
+        "@angular/http": "6.0.6",
+        "@angular/platform-browser": "6.0.6",
+        "@angular/platform-browser-dynamic": "6.0.6",
+        "@angular/router": "6.0.6",
+        "angular-in-memory-web-api": "0.6.0",
         "angular2-uuid": "1.1.1",
+        "ng2-select": "2.0.0",
         "bootstrap": "3.3.7",
-        "core-js": "2.5.0",
+        "core-js": "2.5.7",
         "font-awesome": "4.7.0",
-        "jsplumb": "2.4.3",
-        "moment": "2.18.1",
-        "rxjs": "5.4.3",
-        "zone.js": "0.8.16"
+        "jsplumb": "2.7.9",
+        "moment": "2.22.2",
+        "rxjs": "6.2.1",
+        "rxjs-compat": "6.2.1",
+        "zone.js": "0.8.26"
     },
     "devDependencies": {
-        "@angular/cli": "1.3.1",
-        "@angular/compiler-cli": "4.3.5",
-        "@angular/language-service": "4.3.5",
-        "@types/core-js": "0.9.36",
+        "@angular/cli": "6.0.8",
+        "@angular/compiler-cli": "6.0.6",
+        "@angular/language-service": "6.0.6",
+        "@angular-devkit/build-angular": "0.6.8",
+        "@types/core-js": "2.5.0",
         "@types/jasmine": "2.5.53",
         "@types/jasminewd2": "2.0.2",
-        "@types/node": "8.0.24",
-        "angular2-template-loader": "0.6.2",
-        "codelyzer": "3.1.2",
-        "file-loader": "^0.11.2",
-        "jasmine-core": "2.7.0",
-        "jasmine-spec-reporter": "4.2.1",
-        "karma": "1.7.0",
-        "karma-chrome-launcher": "2.2.0",
-        "karma-cli": "1.0.1",
-        "karma-coverage-istanbul-reporter": "1.3.0",
-        "karma-jasmine": "1.1.0",
-        "karma-jasmine-html-reporter": "0.2.2",
-        "protractor": "5.1.2",
-        "ts-node": "3.3.0",
-        "tslint": "5.6.0",
-        "typedoc": "0.8.0",
-        "typescript": "2.4.2",
-        "webpack": "3.5.6"
+        "@types/node": "10.3.2",
+        "codelyzer": "4.4.2",
+        "protractor": "5.3.2",
+        "ts-node": "5.0.1",
+        "tslint": "5.10.0",
+        "typedoc": "0.11.1",
+        "typescript": "2.7.2",
+        "webpack": "4.8.3"
     }
 }
