@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 Contributors to the Eclipse Foundation
+ * Copyright (c) 2017-2018 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -11,28 +11,29 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  *******************************************************************************/
+import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { RouterModule } from '@angular/router';
 
-import {NgModule} from '@angular/core';
-import {BrowserModule} from '@angular/platform-browser';
-import {RouterModule} from '@angular/router';
+import { AppComponent } from './app.component';
 
-import {AppComponent} from './app.component';
+import { WmContainerComponent } from './components/container/container.component';
+import { WmNodeComponent } from './components/node/node.component';
+import { WmNodeTemplateComponent } from './components/nodetemplate/node-template.component';
+import { WmParameterComponent } from './components/parameter/parameter.component';
+import { WmPropertiesComponent } from './components/property/properties.component';
+import { WmToolbarComponent } from './components/toolbar/toolbar.component';
 
-import {WmContainerComponent} from './components/container/container.component';
-import {WmNodeComponent} from './components/node/node.component';
-import {WmNodeTemplateComponent} from './components/nodetemplate/node-template.component';
-import {WmParameterComponent} from './components/parameter/parameter.component';
-import {WmPropertiesComponent} from './components/property/properties.component';
-import {WmToolbarComponent} from './components/toolbar/toolbar.component';
+import { BroadcastService } from './services/broadcast.service';
+import { JsPlumbService } from './services/jsplumb.service';
+import { ModelService } from './services/model.service';
+import { NodeService } from './services/node.service';
+import { WineryService } from './services/winery.service';
 
-import {BroadcastService} from './services/broadcast.service';
-import {JsPlumbService} from './services/jsplumb.service';
-import {ModelService} from './services/model.service';
-import {NodeService} from './services/node.service';
-import {WineryService} from './services/winery.service';
-
-import {SharedModule} from './shared/shared.module';
-import {HttpService} from './util/http.service';
+import { SharedModule } from './shared/shared.module';
+import { HttpService } from './util/http.service';
+import { HttpClientModule } from '@angular/common/http';
+import { SelectModule } from 'ng2-select';
 
 @NgModule({
     declarations: [
@@ -42,7 +43,7 @@ import {HttpService} from './util/http.service';
         WmNodeTemplateComponent,
         WmParameterComponent,
         WmPropertiesComponent,
-        WmToolbarComponent,
+        WmToolbarComponent
     ],
     providers: [
         BroadcastService,
@@ -56,6 +57,8 @@ import {HttpService} from './util/http.service';
         BrowserModule,
         RouterModule.forRoot([]),
         SharedModule,
+        HttpClientModule,
+        SelectModule,
     ],
     bootstrap: [
         AppComponent,
