@@ -12,12 +12,12 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  *******************************************************************************/
 
-import {Injectable} from '@angular/core';
-import {jsPlumb} from 'jsplumb/dist/js/jsplumb.js';
+import { Injectable } from '@angular/core';
+import { jsPlumb } from 'jsplumb';
 
-import {Node} from '../model/workflow/node';
-import {BroadcastService} from './broadcast.service';
-import {ModelService} from './model.service';
+import { Node } from '../model/workflow/node';
+import { BroadcastService } from './broadcast.service';
+import { ModelService } from './model.service';
 
 /**
  * JsPlumbService
@@ -35,31 +35,31 @@ export class JsPlumbService {
     public connectNode() {
         this.modelService.getNodes()
             .forEach(node => node.connection
-                .forEach(target => this.jsplumbInstance.connect({source: node.id, target})));
+                .forEach(target => this.jsplumbInstance.connect({ source: node.id, target })));
     }
 
     public initJsPlumbInstance() {
-        jsPlumb.ready(() => {
+        jsPlumb.getInstance().ready(() => {
             this.jsplumbInstance = jsPlumb.getInstance();
 
             this.jsplumbInstance.importDefaults({
                 Anchor: ['Top', 'RightMiddle', 'LeftMiddle', 'Bottom'],
                 Connector: [
                     'Flowchart',
-                    {cornerRadius: 0, stub: 0, gap: 3},
+                    { cornerRadius: 0, stub: 0, gap: 3 },
                 ],
                 ConnectionOverlays: [
                     [
                         'Arrow',
-                        {direction: 1, foldback: 1, location: 1, width: 10, length: 10},
+                        { direction: 1, foldback: 1, location: 1, width: 10, length: 10 },
                     ],
-                    ['Label', {label: '', id: 'label', cssClass: 'aLabel'}],
+                    ['Label', { label: '', id: 'label', cssClass: 'aLabel' }],
                 ],
                 connectorPaintStyle: {
                     lineWidth: 2,
                 },
                 Endpoint: 'Blank',
-                PaintStyle: {lineWidth: 1},
+                PaintStyle: { lineWidth: 1 },
             });
 
             this.broadcastService.broadcast(this.broadcastService.jsPlumbInstance,
@@ -71,7 +71,7 @@ export class JsPlumbService {
 
                 info.connection.bind('click', connection => {
                     this.modelService.deleteConnection(connection.sourceId, connection.targetId);
-                    jsPlumb.detach(connection);
+                    this.jsplumbInstance.deleteConnection(connection);
                 });
             });
         });
