@@ -12,12 +12,12 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  *******************************************************************************/
 
-import {Injectable} from '@angular/core';
-import {isNullOrUndefined} from 'util';
+import { Injectable } from '@angular/core';
+import { isNullOrUndefined } from 'util';
 
-import {Node} from '../model/workflow/node';
-import {BroadcastService} from './broadcast.service';
-import {NodeService} from './node.service';
+import { Node } from '../model/workflow/node';
+import { BroadcastService } from './broadcast.service';
+import { NodeService } from './node.service';
 
 /**
  * ModelService
